split out cache locking into separate file
